requirements.txt
fastapi
uvicorn
langchain
langchain-community
pinecone-client
sentence-transformers
huggingface_hub
pypdf
